refactor : 클래스 , 패키지 수정  - ChessDao 클래스 명 -> ChessRepository로 수정  - 중복 ChessRepository 삭제 및 package 수정
