+++ src/main/java/wooteco/chess/repository/JdbcTemplate.java
@@ -1,4 +1,4 @@
-package wooteco.chess.dao;
+package wooteco.chess.repository;
 
 import wooteco.chess.database.MySqlConnector;
