+++ src/main/java/wooteco/chess/repository/SelectJdbcTemplate.java
@@ -1,4 +1,4 @@
-package wooteco.chess.dao;
+package wooteco.chess.repository;
 
 import wooteco.chess.database.MySqlConnector;
 import wooteco.chess.dto.Commands;
