+++ src/test/java/wooteco/chess/database/SpringCommandsDaoTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import wooteco.chess.dao.ChessDao;
+import wooteco.chess.repository.ChessRepository;
 import wooteco.chess.dto.Commands;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -14,30 +14,30 @@ import static org.assertj.core.api.Assertions.assertThat;
 @SpringBootTest
 class SpringCommandsDaoTest {
     @Autowired
-    ChessDao chessDao;
+    ChessRepository chessRepository;
 
     @BeforeEach
     void setUp() {
-        chessDao.save(new Commands("move a2 a4"));
-        chessDao.save(new Commands("move a7 a5"));
+        chessRepository.save(new Commands("move a2 a4"));
+        chessRepository.save(new Commands("move a7 a5"));
     }
 
     @AfterEach
     void afterEach() {
-        chessDao.deleteAll();
+        chessRepository.deleteAll();
     }
 
     @DisplayName("테이블 행 삭제")
     @Test
     void clearCommands() {
-        chessDao.deleteAll();
+        chessRepository.deleteAll();
 
-        assertThat(chessDao.findAll().size()).isZero();
+        assertThat(chessRepository.findAll().size()).isZero();
     }
 
     @DisplayName("전체 행 선택")
     @Test
     void selectCommands() {
-        assertThat(chessDao.findAll().size()).isEqualTo(2);
+        assertThat(chessRepository.findAll().size()).isEqualTo(2);
     }
 }
